sync

+++ spring-context/src/test/java/org/springframework/jmx/export/PropertyPlaceholderConfigurerTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2019 the original author or authors.
+ * Copyright 2002-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,8 +51,7 @@ class PropertyPlaceholderConfigurerTests extends AbstractJmxTests {
 		Integer age = (Integer) getServer().getAttribute(oname, "Age");
 
 		assertThat(name).as("Name is incorrect in JMX").isEqualTo("Rob Harrop");
-		assertThat(age.intValue()).as("Age is incorrect in JMX").isEqualTo(100);
+		assertThat(age).as("Age is incorrect in JMX").isEqualTo(100);
 	}
 
 }
-
