+++ spring-r2dbc/src/main/java/org/springframework/r2dbc/core/UpdatedRowsFetchSpec.java
@@ -30,6 +30,6 @@ public interface UpdatedRowsFetchSpec {
 	 * Get the number of updated rows.
 	 * @return a Mono emitting the number of updated rows
 	 */
-	Mono<Integer> rowsUpdated();
+	Mono<Long> rowsUpdated();
 
 }
