+++ spring-r2dbc/src/main/kotlin/org/springframework/r2dbc/core/UpdatedRowsFetchSpecExtensions.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2020 the original author or authors.
+ * Copyright 2002-2023 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,12 @@
 
 package org.springframework.r2dbc.core
 
-import kotlinx.coroutines.reactive.awaitSingle
+import kotlinx.coroutines.reactor.awaitSingle
 
 /**
  * Coroutines variant of [UpdatedRowsFetchSpec.rowsUpdated].
  *
  * @author Fred Montariol
  */
-suspend fun UpdatedRowsFetchSpec.awaitRowsUpdated(): Int =
+suspend fun UpdatedRowsFetchSpec.awaitRowsUpdated(): Long =
 		rowsUpdated().awaitSingle()
